fix: clean install dir

+++ life/clean.py
@@ -6,5 +6,6 @@ from constants.constants import PATH_MERLIN
 
 
 def clean(context):
-    cmd = 'rm -r %s/*' % PATH_MERLIN
-    logger.info('clean done. [cmd=%s, result=%s]', cmd, os.system(cmd) == 0)
+    if os.getenv("HOME") != os.getcwd():
+        cmd = 'rm -r %s/*' % PATH_MERLIN
+        logger.info('clean done. [cmd=%s, result=%s]', cmd, os.system(cmd) == 0)
